Correct circular dependencies
